refactor: Result 패키지 분리
