+++ src/main/java/blackjack/domain/Result/Results.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.Result;
 
 import java.util.ArrayList;
 import java.util.List;
