+++ src/main/java/blackjack/view/OutputView.java
@@ -1,6 +1,6 @@
 package blackjack.view;
 
-import blackjack.domain.Result;
+import blackjack.domain.Result.Result;
 import blackjack.domain.card.Card;
 import blackjack.domain.player.Participant;
 import blackjack.domain.player.Player;
